+++ src/main/java/vine/vine/repository/NmmainRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import vine.vine.domain.Nmmain;
 
 @Repository
-public interface NmmainRepository extends JpaRepository<Nmmain, Integer> {
+public interface NmmainRepository extends JpaRepository<Nmmain, Long> {
 }
