+++ src/main/java/vine/vine/service/ChargesService.java
@@ -7,5 +7,5 @@ import vine.vine.domain.dto.response.ChargesResponse;
 public interface ChargesService {
     Page<ChargesResponse> getAllCharges(Pageable pageable);
     String getPrisonerCharges();
-    String getPrisonerChargesByBookingId(Integer bookId);
+    String getPrisonerChargesByBookingId(Long bookId);
 }
